siproxd bug 3086351 plugin_shortdial.c:139: warning: comparison is always false due to limited range of data type

+++ src/plugin_shortdial.c
@@ -136,7 +136,7 @@ static int plugin_shortdial(sip_ticket_t *ticket) {
       return STS_SUCCESS; /* ignore */
 
    shortcut_no = atoi(&(req_url->username[1]));
-   if ((shortcut_no <= 0) || (shortcut_no >= LONG_MAX)) return STS_SUCCESS; /* not a number */
+   if ((shortcut_no <= 0) || (shortcut_no >= INT_MAX)) return STS_SUCCESS; /* not a number */
 
    /* requested number is not defined (out of range) */
    if (shortcut_no > plugin_cfg.shortdial_entry.used) {
